requirements.txt
requests 
beautifulsoup4
numpy
selenium
html5lib
webdriver-manager
pyyaml 
pandas
psycopg2-binary==2.9.10
